+++ bevy_asset_loader/src/asset_loader/dynamic_asset_systems.rs
@@ -1,67 +1,13 @@
 #[cfg(feature = "dynamic_assets")]
-use bevy::asset::AssetServer;
-use bevy::asset::{Assets, LoadState};
+use crate::asset_loader::dynamic_asset::DynamicAssetCollection;
 #[cfg(feature = "dynamic_assets")]
-use bevy::ecs::prelude::World;
+use crate::asset_loader::{AssetLoaderConfiguration, DynamicAssets, LoadingState};
 #[cfg(feature = "dynamic_assets")]
-use bevy::ecs::schedule::{State, StateData};
+use bevy::asset::LoadState;
 #[cfg(feature = "dynamic_assets")]
-use bevy::utils::HashMap;
-
+use bevy::ecs::schedule::StateData;
 #[cfg(feature = "dynamic_assets")]
-use bevy::reflect::TypeUuid;
-
-#[cfg(feature = "dynamic_assets")]
-use crate::{AssetLoaderConfiguration, DynamicAssets, LoadingState};
-
-/// These asset variants can be loaded from configuration files. They will then replace
-/// a dynamic asset based on their keys.
-#[cfg_attr(feature = "dynamic_assets", derive(serde::Deserialize))]
-pub enum DynamicAsset {
-    /// A dynamic asset directly loaded from a single file
-    #[cfg_attr(feature = "dynamic_assets", serde(alias = "Folder"))]
-    File {
-        /// Asset file path
-        path: String,
-    },
-    /// A dynamic standard material asset directly loaded from an image file
-    #[cfg(feature = "render")]
-    StandardMaterial {
-        /// Asset file path
-        path: String,
-    },
-    /// A dynamic texture atlas asset loaded from a sprite sheet
-    #[cfg(feature = "render")]
-    TextureAtlas {
-        /// Asset file path
-        path: String,
-        /// The image width in pixels
-        tile_size_x: f32,
-        /// The image height in pixels
-        tile_size_y: f32,
-        /// Columns on the sprite sheet
-        columns: usize,
-        /// Rows on the sprite sheet
-        rows: usize,
-        /// Padding between columns in pixels
-        padding_x: Option<f32>,
-        /// Padding between rows in pixels
-        padding_y: Option<f32>,
-    },
-}
-
-impl DynamicAsset {
-    /// Path to the asset file of the dynamic asset
-    pub fn get_file_path(&self) -> &str {
-        match self {
-            DynamicAsset::File { path } => path,
-            #[cfg(feature = "render")]
-            DynamicAsset::StandardMaterial { path } => path,
-            #[cfg(feature = "render")]
-            DynamicAsset::TextureAtlas { path, .. } => path,
-        }
-    }
-}
+use bevy::prelude::{AssetServer, Assets, State, World};
 
 #[cfg(feature = "dynamic_assets")]
 pub(crate) fn load_dynamic_asset_collections<S: StateData>(world: &mut World) {
@@ -79,7 +25,9 @@ pub(crate) fn load_dynamic_asset_collections<S: StateData>(world: &mut World) {
 
     let files = asset_loader_config.get_asset_collection_files(state.current());
     if files.is_empty() {
-        loading_state.set(LoadingState::LoadingAssets);
+        loading_state
+            .set(LoadingState::LoadingAssets)
+            .expect("Failed to set loading state");
         return;
     }
     for file in files {
@@ -88,11 +36,6 @@ pub(crate) fn load_dynamic_asset_collections<S: StateData>(world: &mut World) {
             .push(asset_server.load(&file));
     }
 }
-
-#[derive(serde::Deserialize, TypeUuid)]
-#[uuid = "2df82c01-9c71-4aa8-adc4-71c5824768f1"]
-#[cfg(feature = "dynamic_assets")]
-pub struct DynamicAssetCollection(pub(crate) HashMap<String, DynamicAsset>);
 
 #[cfg(feature = "dynamic_assets")]
 pub(crate) fn check_dynamic_asset_collections<S: StateData>(world: &mut World) {
